crypto: Refactor and split out the gossiping logic
